refactor: change currency input to generic onChange function feature
