--- src/components/Form/Input/interface.d.ts
@@ -1,7 +1,0 @@
-import { InputHTMLAttributes } from "react";
-
-interface IInputProps extends InputHTMLAttributes<HTMLInputElement> {
-  label?: string;
-  type?: HTMLInputTypeAttribute | "currency";
-  error?: string;
-}
