--- src/components/Form/index.tsx
@@ -1,1 +1,0 @@
-export { CurrencyInput } from "./CurrencyInput";
