+++ src/components/FormFields/utils/interface.d.ts
@@ -1,0 +1,7 @@
+import { InputHTMLAttributes } from "react";
+
+interface IInputBaseProps extends InputHTMLAttributes<HTMLInputElement> {
+  label?: string;
+  name: string;
+  options?: RegisterOptions;
+}
